made necessary confguration and code changes to get it to deploy via github actions.
Note that I had to copy the DomainTestFactory into the presentation layer tests (duplication) since the pom configuration in that project removed the domain layer jar out of scope and thus the domain objects could no longer be found in the main package.

The way to share the DomainTestFactory is to have it in another project and share it this way.